* all from with-tracing branch

+++ Server/clean.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 
-# delete all binary files
-find . -type f -executable -exec sh -c "file -i '{}' | grep -q 'x-executable; charset=binary'" \; -print | xargs rm -f
-
 rm -f *.log
-rm -rf ./build
-rm -rf ./data
+rm -rf ./build ./Client ./data
 rm -rf ./storage/data
-rm -rf /var/tmp/n3-privacy/meta/
-rm -rf /var/tmp/n3-privacy/badger/
+rm -f ./config/copy.toml
+rm -f *.toml
+
+if [ $# -gt 0 ]; then
+  if [ $1 = "rmdb" ]; then
+    rm -rf /var/tmp/n3-privacy/meta/
+    rm -rf /var/tmp/n3-privacy/badger/
+    echo "database is deleted"
+  else
+    echo "[$1] is invalid argument, ignored"
+  fi
+fi
